+++ scripts/session_restorer.sh
@@ -71,7 +71,7 @@ new_pane() {
 restore_pane() {
 	local pane="$1"
 	echo "$pane" |
-	while IFS=$'\t' read line_type session_name window_number window_name dir; do
+	while IFS=$'\t' read line_type session_name window_number dir window_name; do
 		if window_exists "$session_name" "$window_number"; then
 			new_pane "$session_name" "$window_number" "$window_name" "$dir"
 		elif session_exists "$session_name"; then