+++ scripts/session_saver.sh
@@ -13,9 +13,9 @@ pane_format() {
 	format+="${delimiter}"
 	format+="#{window_index}"
 	format+="${delimiter}"
-	format+="#{window_name}"
-	format+="${delimiter}"
 	format+="#{pane_current_path}"
+	format+="${delimiter}"
+	format+="#{window_name}"
 	echo "$format"
 }