+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### master
+
+### v0.0.3, 2014-08-26
 - save and restore current and alternate session
 - fix a bug with non-existing window names
 - restore active pane for each window that has multiple panes